Editando e Excluindo Registros

+++ src/main/java/com/br/AdMon/dao/ContaDao.java
@@ -1,9 +1,11 @@
 package com.br.AdMon.dao;
 
+import java.math.BigInteger;
+
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.br.AdMon.models.Contas;
 
-public interface ContaDao extends JpaRepository<Contas, Integer>{
+public interface ContaDao extends JpaRepository<Contas, BigInteger>{
 
 }
